private api integeration made successfully
