+++ app/api/join-by-invite/route.ts
@@ -13,7 +13,12 @@ export async function PATCH(req: NextRequest) {
     if (!habit) {
       return NextResponse.json({ message: 'Habit not found' }, { status: 404 })
     }
-
+    if (habit.participants.includes(userid)) {
+      return NextResponse.json(
+        { message: "User is already a participant in this habit" },
+        { status: 400 }
+      );
+    }
     habit.participants.push(userid)
     await habit.save()
